simple lifetime and example of lifetime that wont work
